+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.4.0
+  * support credential deletion
+  * Read / Delete and Enumerate operations throw dedicated Exception: **CredentialManagerException**
+
 # 0.3.0
   * expose Credential::Secret as a byte[] instead of a string. Up to api client to decode content